feat(core): allow switch workspace in loading fallback (#6129)
