+++ packages/frontend/core/src/components/root-app-sidebar/index.tsx
@@ -1,5 +1,4 @@
 import { AnimatedDeleteIcon } from '@affine/component';
-import { Menu } from '@affine/component/ui/menu';
 import { useAsyncCallback } from '@affine/core/hooks/affine-async-hooks';
 import { CollectionService } from '@affine/core/modules/collection';
 import { apis, events } from '@affine/electron-api';
@@ -8,12 +7,11 @@ import { FolderIcon, SettingsIcon } from '@blocksuite/icons';
 import { type Doc } from '@blocksuite/store';
 import { useDroppable } from '@dnd-kit/core';
 import { useLiveData, useService, type Workspace } from '@toeverything/infra';
-import { useAtom, useAtomValue } from 'jotai';
+import { useAtomValue } from 'jotai';
 import { nanoid } from 'nanoid';
 import type { HTMLAttributes, ReactElement } from 'react';
-import { forwardRef, Suspense, useCallback, useEffect } from 'react';
+import { forwardRef, useCallback, useEffect } from 'react';
 
-import { openWorkspaceListModalAtom } from '../../atoms';
 import { useAppSettingHelper } from '../../hooks/affine/use-app-setting-helper';
 import { useDeleteCollectionInfo } from '../../hooks/affine/use-delete-collection-info';
 import { getDropItemId } from '../../hooks/affine/use-sidebar-drag';
@@ -41,8 +39,7 @@ import { CollectionsList } from '../pure/workspace-slider-bar/collections';
 import { AddCollectionButton } from '../pure/workspace-slider-bar/collections/add-collection-button';
 import { AddFavouriteButton } from '../pure/workspace-slider-bar/favorite/add-favourite-button';
 import FavoriteList from '../pure/workspace-slider-bar/favorite/favorite-list';
-import { UserWithWorkspaceList } from '../pure/workspace-slider-bar/user-with-workspace-list';
-import { WorkspaceCard } from '../pure/workspace-slider-bar/workspace-card';
+import { WorkspaceSelector } from '../workspace-selector';
 import ImportPage from './import-page';
 import { AppSidebarJournalButton } from './journal-button';
 import { UpdaterButton } from './updater-button';
@@ -102,9 +99,6 @@ export const RootAppSidebar = ({
   const { appSettings } = useAppSettingHelper();
   const docCollection = currentWorkspace.docCollection;
   const t = useAFFiNEI18N();
-  const [openUserWorkspaceList, setOpenUserWorkspaceList] = useAtom(
-    openWorkspaceListModalAtom
-  );
   const currentPath = useLiveData(useService(Workbench).location).pathname;
 
   const onClickNewPage = useAsyncCallback(async () => {
@@ -149,10 +143,6 @@ export const RootAppSidebar = ({
   const trashDroppable = useDroppable({
     id: dropItemId,
   });
-  const closeUserWorkspaceList = useCallback(() => {
-    setOpenUserWorkspaceList(false);
-  }, [setOpenUserWorkspaceList]);
-  const userInfo = useDeleteCollectionInfo();
 
   const collection = useService(CollectionService);
   const { node, open } = useEditCollectionName({
@@ -170,6 +160,7 @@ export const RootAppSidebar = ({
         console.error(err);
       });
   }, [docCollection.id, collection, navigateHelper, open]);
+  const userInfo = useDeleteCollectionInfo();
 
   const allPageActive = currentPath === '/all';
 
@@ -188,31 +179,7 @@ export const RootAppSidebar = ({
         titles={deletePageTitles}
       />
       <SidebarContainer>
-        <Menu
-          rootOptions={{
-            open: openUserWorkspaceList,
-          }}
-          items={
-            <Suspense>
-              <UserWithWorkspaceList onEventEnd={closeUserWorkspaceList} />
-            </Suspense>
-          }
-          contentOptions={{
-            // hide trigger
-            sideOffset: -58,
-            onInteractOutside: closeUserWorkspaceList,
-            onEscapeKeyDown: closeUserWorkspaceList,
-            style: {
-              width: '300px',
-            },
-          }}
-        >
-          <WorkspaceCard
-            onClick={useCallback(() => {
-              setOpenUserWorkspaceList(true);
-            }, [setOpenUserWorkspaceList])}
-          />
-        </Menu>
+        <WorkspaceSelector />
         <QuickSearchInput
           data-testid="slider-bar-quick-search-button"
           onClick={onOpenQuickSearchModal}
